chore(configuration): Renamed ConfigurationStatus
