+++ src/components/common/configuration/ConfigurationStatus.vue
