+++ src/components/data-workflows/gcs-to-gcs/GcsToGcsConfs.vue
@@ -157,7 +157,7 @@ import { mapGetters } from 'vuex';
 import VueJsonPretty from 'vue-json-pretty';
 import store from '@/store';
 import _ from 'lodash';
-import ActivatedStatusChip from '../../common/chips/ActivatedStatusChip.vue';
+import ActivatedStatusChip from '../../common/configuration/ConfigurationStatus.vue';
 import ConfsComponent from '@/mixins/confsComponent.js';
 import DataManagementFilters from '../../common/DataManagementFilters';
