+++ src/components/data-workflows/tables-to-tables/TablesToTablesConfs.vue
@@ -142,7 +142,7 @@ import VueJsonPretty from 'vue-json-pretty';
 import store from '@/store';
 import _ from 'lodash';
 import DataManagementFilters from '../../common/DataManagementFilters';
-import ActivatedStatusChip from '../../common/chips/ActivatedStatusChip.vue';
+import ActivatedStatusChip from '../../common/configuration/ConfigurationStatus.vue';
 import ConfsComponent from '@/mixins/confsComponent.js';
 
 export default {
